00_SIGNALduino.pm:
Added new parsing sub for messages with sync. They now come in unfiltered beginning with MS...
Identification is not completly in this module.

Therefore some parameters for the protocols are added:
length_min and length_max . If set, messages must have number of bits within this defined range.

Also prepared a function to split off the messages for all parsers to reduce code doubling